@@ -11,13 +11,13 @@
     ],
     "minimum-stability": "dev",
     "require": {
-        "php": "^8.1",
-        "illuminate/support": ">=9.0",
-        "illuminate/console": ">=9.0",
-        "antonioprimera/laravel-generator-command": "^2.0"
+        "php": "^8.2",
+        "illuminate/support": ">=11.0",
+        "illuminate/console": ">=11.0",
+        "antonioprimera/laravel-generator-command": "^2.4"
     },
     "require-dev": {
-        "orchestra/testbench": "^7"
+        "orchestra/testbench": "^9"
     },
     "extra": {
         "laravel": {
