@@ -3,22 +3,23 @@ namespace AntonioPrimera\Bapi\Tests\Unit;
 
 use AntonioPrimera\Bapi\Tests\TestContext\CallTestBapi;
 use AntonioPrimera\Bapi\Tests\TestContext\TestModel;
+use PHPUnit\Framework\Attributes\Test;
 
 class BapiCallTest extends \Orchestra\Testbench\TestCase
 {
-	/** @test */
+	#[Test]
 	public function a_bapi_can_be_called_statically_by_its_run_method()
 	{
 		$this->assertTrue(is_array(CallTestBapi::run(testModel: new TestModel(), num: 12)));
 	}
 	
-	/** @test */
+	#[Test]
 	public function a_bapi_can_be_instantiated_and_the_run_method_can_be_called()
 	{
 		$this->assertTrue(is_array((new CallTestBapi())->run(testModel: new TestModel(), num: 12)));
 	}
 	
-	/** @test */
+	#[Test]
 	public function a_bapi_can_be_invoked()
 	{
 		$bapi = new CallTestBapi();
@@ -26,7 +27,7 @@ class BapiCallTest extends \Orchestra\Testbench\TestCase
 		$this->assertIsArray($bapi(testModel: new TestModel(), num: 12));
 	}
 	
-	/** @test */
+	#[Test]
 	public function the_handle_method_is_run_when_the_bapi_is_run()
 	{
 		$bapi = new CallTestBapi();
@@ -36,7 +37,7 @@ class BapiCallTest extends \Orchestra\Testbench\TestCase
 		$this->assertTrue($bapi->handleMethodWasCalled);
 	}
 	
-	/** @test */
+	#[Test]
 	public function on_running_the_bapi_the_arguments_provided_are_available_to_all_methods_in_the_run_lifecycle()
 	{
 		$bapi = new CallTestBapi();
@@ -62,7 +63,7 @@ class BapiCallTest extends \Orchestra\Testbench\TestCase
 		$this->assertArraysAreIdentical($expectedAttributes, $bapi->attributeLists['processResult']);
 	}
 	
-	/** @test */
+	#[Test]
 	public function the_bapi_has_a_strict_lifecycle_calling_all_hooks()
 	{
 		$bapi = new CallTestBapi();
@@ -79,7 +80,7 @@ class BapiCallTest extends \Orchestra\Testbench\TestCase
 		$this->assertEquals('processResult', 		$bapi->methodCalls[3]);
 	}
 	
-	/** @test */
+	#[Test]
 	public function the_after_handle_hook_can_change_the_result_of_the_handle_method_before_returning_the_bapi_result()
 	{
 		$testModel = new TestModel();
